+++ libs/react-components/src/tedi/components/layout/sidenav/components/sidenav-mobile/sidenav-mobile.tsx
@@ -21,6 +21,8 @@ type SideNavMobileProps<C extends React.ElementType = 'a'> = {
   isOpen: boolean;
   onClose: () => void;
   showOverlay?: boolean;
+  id?: string;
+  className?: string;
 };
 
 export const SideNavMobile = <C extends React.ElementType = 'a'>({
@@ -30,6 +32,8 @@ export const SideNavMobile = <C extends React.ElementType = 'a'>({
   isOpen,
   onClose,
   showOverlay = true,
+  id,
+  className,
 }: SideNavMobileProps<C>) => {
   const { getLabel } = useLabels();
   const [navigationStack, setNavigationStack] = useState<NavigationLevel<C>[]>([{ items: navItems }]);
@@ -143,8 +147,9 @@ export const SideNavMobile = <C extends React.ElementType = 'a'>({
   const content = (
     <nav
       data-name="mobile-sidenav"
-      className={classNames(styles['tedi-sidenav'], styles['tedi-sidenav--mobile'])}
+      className={classNames(styles['tedi-sidenav'], styles['tedi-sidenav--mobile'], className)}
       aria-label={ariaLabel}
+      id={id}
     >
       <div className={styles['tedi-sidenav__list']}>
         {!isRootLevel && (